# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vasooli API",
    "description": "AI-first revenue recovery — one policy layer, one audit trail, three engines.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/audit/batches/{batch_id}/summary": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Batch Summary",
        "description": "The headline metrics for one batch, recomputed from its entries.\n\nRecomputed rather than read from the batch's stored snapshot, so the number\nreported here and the trail behind it cannot disagree.",
        "operationId": "batch_summary_api_v1_audit_batches__batch_id__summary_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/receivables/runs": {
      "post": {
        "tags": [
          "receivables"
        ],
        "summary": "Trigger Run",
        "description": "Run the full loop over an invoice ledger and return the honest summary.",
        "operationId": "trigger_run_api_v1_receivables_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__api__v1__routes__receivables__RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__engines__receivables__schemas__RunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "receivables"
        ],
        "summary": "List Runs",
        "description": "Every receivables run, newest first. Each carries its seed.",
        "operationId": "list_runs_api_v1_receivables_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchRunRead"
                  },
                  "title": "Response List Runs Api V1 Receivables Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receivables/runs/{batch_id}": {
      "get": {
        "tags": [
          "receivables"
        ],
        "summary": "Get Run",
        "description": "One run's record, including the dataset batch id it consumed.",
        "operationId": "get_run_api_v1_receivables_runs__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receivables/runs/{batch_id}/summary": {
      "get": {
        "tags": [
          "receivables"
        ],
        "summary": "Get Run Summary",
        "description": "The run's own summary — the engine-specific figures the dashboard reads.\n\nServed from what the run stored rather than recomputed, because the splits in\nit (per-ageing-bucket recovery, the promise register, extraction accuracy)\nneed the run's outcomes and the dataset behind them, and nothing serving an\nAPI request should be re-running an engine.\n\nThe money figures inside it were read off the audit trail at the end of the\nrun, and `/audit/batches/{batch_id}/summary` recomputes those independently\n— so the stored copy is checkable against the trail rather than merely\nconvenient. Quote that route for a headline; quote this one for the breakdown.",
        "operationId": "get_run_summary_api_v1_receivables_runs__batch_id__summary_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__engines__receivables__schemas__RunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receivables/runs/{batch_id}/extraction": {
      "get": {
        "tags": [
          "receivables"
        ],
        "summary": "Get Extraction Score",
        "description": "Extraction accuracy for one run, with its confusion matrix and failures.\n\nServed from the run's stored summary rather than recomputed, because the\ncomparison needs the manifest and the manifest is not something an API\nrequest should be reaching for. The numbers are the ones the run reported.",
        "operationId": "get_extraction_score_api_v1_receivables_runs__batch_id__extraction_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionScore"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receivables/runs/{batch_id}/worklist": {
      "get": {
        "tags": [
          "receivables"
        ],
        "summary": "Get Worklist",
        "description": "The ranked worklist, highest priority first, with every score breakdown.\n\nDeprioritised invoices stay on the list rather than being filtered out: \"we\nchose not to chase this, and here is the rule\" is the claim worth making, and\nfiltering would make the same decision invisible.",
        "operationId": "get_worklist_api_v1_receivables_runs__batch_id__worklist_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "deprioritised",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only invoices held back by a live promise (policy-bounds:PR2).",
              "title": "Deprioritised"
            },
            "description": "Only invoices held back by a live promise (policy-bounds:PR2)."
          },
          {
            "name": "ageing_bucket",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ageing Bucket"
            }
          },
          {
            "name": "next_step",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Next Step"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceChaseStateRead"
                  },
                  "title": "Response Get Worklist Api V1 Receivables Runs  Batch Id  Worklist Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receivables/runs/{batch_id}/promises": {
      "get": {
        "tags": [
          "receivables"
        ],
        "summary": "List Promises",
        "description": "The promise register: every commitment extracted, and what became of it.",
        "operationId": "list_promises_api_v1_receivables_runs__batch_id__promises_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`active`, `kept`, `broken` or `superseded`.",
              "title": "Status"
            },
            "description": "`active`, `kept`, `broken` or `superseded`."
          },
          {
            "name": "conditional",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Conditional commitments are tracked apart from firm ones (policy-bounds:PP2).",
              "title": "Conditional"
            },
            "description": "Conditional commitments are tracked apart from firm ones (policy-bounds:PP2)."
          },
          {
            "name": "customer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PromiseToPayRead"
                  },
                  "title": "Response List Promises Api V1 Receivables Runs  Batch Id  Promises Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receivables/runs/{batch_id}/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "receivables"
        ],
        "summary": "Get Invoice Timeline",
        "description": "One invoice's full timeline: contacts, reply, reading, promise, decisions.\n\nComplete on purpose — this is the demo surface. Every rule behind every step\nis in the response, including the ones that refused.",
        "operationId": "get_invoice_timeline_api_v1_receivables_runs__batch_id__invoices__invoice_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceTimeline"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receivables/runs/{batch_id}/communications": {
      "get": {
        "tags": [
          "receivables"
        ],
        "summary": "List Communications",
        "description": "Every reminder the run drafted. **None were sent** — see ADR 0009.",
        "operationId": "list_communications_api_v1_receivables_runs__batch_id__communications_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`marked_for_delivery`, `held` or `suppressed`. Never `sent`: nothing here dispatches.",
              "title": "Status"
            },
            "description": "`marked_for_delivery`, `held` or `suppressed`. Never `sent`: nothing here dispatches."
          },
          {
            "name": "rung",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rung"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceCommunicationRead"
                  },
                  "title": "Response List Communications Api V1 Receivables Runs  Batch Id  Communications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchSummary": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "entries": {
            "type": "integer",
            "title": "Entries"
          },
          "amount_at_risk_paise": {
            "type": "integer",
            "title": "Amount At Risk Paise"
          },
          "amount_recovered_paise": {
            "type": "integer",
            "title": "Amount Recovered Paise"
          },
          "recovery_rate": {
            "type": "number",
            "title": "Recovery Rate"
          },
          "action_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Action Counts"
          },
          "outcome_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Outcome Counts"
          },
          "engine_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Engine Counts"
          },
          "by_source": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceBreakdown"
            },
            "type": "object",
            "title": "By Source"
          },
          "policy_violations": {
            "type": "integer",
            "title": "Policy Violations"
          },
          "blocked_count": {
            "type": "integer",
            "title": "Blocked Count"
          },
          "halted_count": {
            "type": "integer",
            "title": "Halted Count"
          },
          "escalated_count": {
            "type": "integer",
            "title": "Escalated Count"
          },
          "abstained_count": {
            "type": "integer",
            "title": "Abstained Count"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "entries",
          "amount_at_risk_paise",
          "amount_recovered_paise",
          "recovery_rate",
          "action_counts",
          "outcome_counts",
          "engine_counts",
          "by_source",
          "policy_violations",
          "blocked_count",
          "halted_count",
          "escalated_count",
          "abstained_count"
        ],
        "title": "BatchSummary",
        "description": "The headline number, computed honestly from the log.\n\nNever from a separately maintained counter — a counter can drift from the\ntrail it claims to summarise, and then the demo's headline figure and its\nevidence disagree."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}